feat(editor): always-on WYSIWYG editor with autosave, rename coeditor -> editor
Removes the read/edit toggle in FileView so the WYSIWYG Coeditor is always
mounted for the current doc version (DiffView still owns history), replacing
the explicit Save/Done button with idle + hard-cap autosave timers in
useCoeditSession. Renames lib/coeditor/ to lib/editor/ and moves
CoeditPresenceBar there, since it's now the sole owner of doc rendering,
collab, presence, and comments. Replaces the DOM/react-markdown <-> raw-offset
comment-anchoring bridge (commentAnchor.ts, rehypeSourcePos.ts) with
CodeMirror-native offsets and decorations (lib/editor/comments.ts), since
CodeMirror's doc is already the raw string. Renaming moves to DocTitle via
Opal's Content editable/onTitleChange, decoupled from the coedit session.
